Functions for updating existing pkl and csv files.

+++ activity_tracker.py
@@ -38,7 +38,7 @@ def retrieve_activity(name: str) -> tuple:
 def activity_to_df(comments, posts):
     """Function to create a dataframe from the praw comments and posts sublistings."""
     activitydata = {
-        "timestamp": [],
+        "timestamp (UTC)": [],
         "id": [],
         "karma": [],
         "subreddit": [],
@@ -47,7 +47,7 @@ def activity_to_df(comments, posts):
         "type": [],
     }
     for comment in comments:
-        activitydata["timestamp"].append(comment.created_utc)
+        activitydata["timestamp (UTC)"].append(comment.created_utc)
         activitydata["id"].append(comment.id)
         activitydata["karma"].append(comment.score)
         activitydata["subreddit"].append(comment.subreddit.display_name)
@@ -55,7 +55,7 @@ def activity_to_df(comments, posts):
         activitydata["account_name"].append(comment.author.name)
         activitydata["type"].append("Comment")
     for post in posts:
-        activitydata["timestamp"].append(post.created_utc)
+        activitydata["timestamp (UTC)"].append(post.created_utc)
         activitydata["id"].append(post.id)
         activitydata["karma"].append(post.score)
         activitydata["subreddit"].append(post.subreddit.display_name)
@@ -68,18 +68,62 @@ def activity_to_df(comments, posts):
     )
     return activitydf
 
+
 def save_to_pickle(activitydf: pd.DataFrame, karmadf: pd.DataFrame):
     """Save given pandas dataframe args to a pickle file in data directory."""
-    activitydf.to_pickle(Path.cwd() / "data" / "activity_data.pkl")
-    karmadf.to_pickle(Path.cwd() / "data" / "karmatracker.pkl")
+    check = input(
+        "Are you sure you would like to save the pkl files? (overwrites existing files) (Y/N)"
+    )
+    if check == "Y":
+        activitydf.to_pickle(Path.cwd() / "data" / "activity_data.pkl")
+        karmadf.to_pickle(Path.cwd() / "data" / "karmatracker.pkl")
+    else:
+        return
+
 
 def save_to_csv(activitydf: pd.DataFrame, karmadf: pd.DataFrame):
     """Save given pandas dataframe args to a csv file in data directory."""
-    activitydf.to_csv(Path.cwd() / "data" / "activity_data.csv")
-    karmadf.to_csv(Path.cwd() / "data" / "karmatracker.csv")
+    check = input(
+        "Are you sure you would like to save the csv files? (overwrites existing files) (Y/N)"
+    )
+    if check == "Y":
+        activitydf.to_csv(Path.cwd() / "data" / "activity_data.csv")
+        karmadf.to_csv(Path.cwd() / "data" / "karmatracker.csv")
+    else:
+        return
+
+
+def update_pickle(activitydf: pd.DataFrame, karmadf: pd.DataFrame):
+    """Update existing pickle files with new dataframes."""
+    activitydf_0 = pd.read_pickle(Path.cwd() / "data" / "activity_data.pkl")
+    karmadf_0 = pd.read_pickle(Path.cwd() / "data" / "karmatracker.pkl")
+    activitydf_0.append(activitydf, ignore_index=True)
+    karmadf_0.append(karmadf, ignore_index=True)
+    check = input("Are you sure you would like to update the pkl files? (Y/N)")
+    if check == "Y":
+        activitydf_0.to_pickle(Path.cwd() / "data" / "activity_data.pkl")
+        karmadf_0.to_pickle(Path.cwd() / "data" / "karmatracker.pkl")
+    else:
+        return
+
+
+def update_csv(activitydf: pd.DataFrame, karmadf: pd.DataFrame):
+    """Update exisiting csv files with new dataframes."""
+    activitydf_0 = pd.read_csv(Path.cwd() / "data" / "activity_data.csv", index_col=0)
+    karmadf_0 = pd.read_csv(Path.cwd() / "data" / "karmatracker.csv", index_col=0)
+    activitydf_0 = activitydf_0.append(activitydf, ignore_index=True)
+    karma_df_0 = karmadf_0.append(karmadf, ignore_index=True)
+    check = input("Are you sure you would like to update the csv files? (Y/N)")
+    if check == "Y":
+        activitydf_0.to_csv(Path.cwd() / "data" / "activity_data.csv")
+        karmadf_0.to_csv(Path.cwd() / "data" / "karmatracker.csv")
+    else:
+        return
+
 
 if __name__ == "__main__":
-    accounts = ["mohiemen", "KingPZe"]
+    # accounts = ["mohiemen", "KingPZe"]
+    accounts = ["orchid_breeder", "DurovCode"]
     user_info = []
     for account in accounts:
         user = retrieve_account(account)
@@ -97,5 +141,9 @@ if __name__ == "__main__":
     print(karmadf)
 
     # Only uncomment these commands when you are first saving or you want to overwrite old files.
-    save_to_pickle(activitydf,karmadf)
-    save_to_csv(activitydf,karmadf)
+    # save_to_pickle(activitydf, karmadf)
+    # save_to_csv(activitydf, karmadf)
+
+    # Functions to update existing files.
+    # update_pickle(activitydf, karmadf)
+    # update_csv(activitydf, karmadf)
